### 2018-09-16 #951 @whedon
[PRE REVIEW]: dyPolyChord: dynamic nested sampling with PolyChord
**Submitting author:** @ejhigson (<a href="http://orcid.org/0000-0001-8383-4614">Edward Higson</a>)
**Repository:** <a href="https://github.com/ejhigson/dyPolyChord" target ="_blank">https://github.com/ejhigson/dyPolyChord</a>
**Version:** v0.0.4
