### 2018-09-16 #951 @arfon
> Note that while dyPolyChord is fully open source, one of its dependencies (PolyChord) does not fully satisfy the OSI open source criteria but is freely available for academic use and is widely used in research (it can be downloaded at https://ccpforge.cse.rl.ac.uk/gf/project/polychord/). I think dyPolyChord is still eligible for JOSS, but let me know if I am mistaken.

👋 @jakevdp - the submitting author suggested you as the handling editor. In addition, they posted the above with the submission. From a quick inspection, it looks like this submission does satisfy the JOSS requirements.
