### 2018-09-16 83ba43b
minor references tidying and fix
