+++ docs/paper/paper.bib
@@ -193,21 +193,14 @@ year = {2011}
 }
 @misc{Jones2001,
 author = {Jones, Eric and Oliphant, Travis and Peterson, Pearu and Others},
-title = {{{\{}SciPy{\}}: Open source scientific tools for {\{}Python{\}}}},
+title = {SciPy: Open source scientific tools for Python},
 url = {http://www.scipy.org/},
 year = {2001}
 }
-@article{Oliphant2010,
-abstract = {Numpy guide},
-archivePrefix = {arXiv},
-arxivId = {arXiv:1011.1669v3},
-author = {Oliphant, Travis E},
-doi = {10.1016/j.jmoldx.2015.02.001},
-eprint = {arXiv:1011.1669v3},
-journal = {Methods},
-pages = {378},
-pmid = {25908243},
-title = {{Guide to NumPy}},
-volume = {1},
-year = {2010}
+@book{Oliphant2006,
+title={A guide to NumPy},
+author={Oliphant, Travis E},
+volume={1},
+year={2006},
+publisher={Trelgol Publishing USA}
 }
