### 2018-09-17 #951 @arfon
@ejhigson - could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
