### 2018-09-17 #951 @andrewfowlie
Dear all,

This is a cool way to find reviewers and perform peer review. I am very willing to review the work and participate in the JOSS journal. You can contact me here or by email.

Andrew
