### 2018-09-19 fe86d1c
minor update to paper.md to include discussion of the computational overhead of allocating samples via dynamic nested sampling
