### 2018-09-19 #951 @ejhigson
(minor addition to paper)
