### 2018-09-20 1ff3aad
added a couple more sentences about performance
