### 2018-09-20 #965 @ejhigson
(minor addition to paper.md giving more details about performance)
