### 2018-09-21 ejhigson/dyPolyChord#1 @andrewfowlie
Documentation - statement of need
Hi there,

For my [review](https://github.com/openjournals/joss-reviews/issues/965),

> ### Documentation
>
> - [ ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I couldn't see this information clearly in the [docs](https://dypolychord.readthedocs.io/en/latest/index.html). It was clear in the software paper, though.
