### 2018-09-21 ejhigson/dyPolyChord#2 @andrewfowlie
Functionality - performance
Hi there,

For my [review](https://github.com/openjournals/joss-reviews/issues/965), I am trying to check off this requirement,

> - [ ] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

How can I verify the performance gains achieved by your code? Is there a particular example that illustrates them?
