### 2018-09-22 684db65
adding performance section of docs following paper
