+++ docs/index.rst
@@ -14,11 +14,14 @@ dyPolyChord
 .. image:: https://img.shields.io/badge/license-MIT-blue.svg
    :target: https://github.com/ejhigson/dyPolyChord/blob/master/LICENSE
 
-``dyPolyChord`` uses PolyChord to perform dynamic nested sampling by running it from within a python wrapper. Any likelihoods and priors which work with PolyChord can be used (python, C++ or Fortran), and the output files produced are in the PolyChord format.
-For more details of dynamic nested sampling see the dynamic nested sampling paper (`Higson et al., 2017 <https://arxiv.org/abs/1704.03459>`_).
+``dyPolyChord`` implements dynamic nested sampling using the efficient ``PolyChord`` sampler to provide state-of-the-art nested sampling performance.
+Any likelihoods and priors which work with ``PolyChord`` can be used (Python, C++ or Fortran), and the output files produced are in the ``PolyChord`` format.
 
 To get started, see the `installation instructions <http://dyPolyChord.readthedocs.io/en/latest/install.html>`_ and the `demo <http://dyPolyChord.readthedocs.io/en/latest/demo.html>`_.
-N.B. ``dyPolyChord`` requires PolyChord v1.14 or higher.
+N.B. ``dyPolyChord`` requires ``PolyChord`` v1.14 or higher.
+
+For more details about dynamic nested sampling, see the dynamic nested sampling paper (`Higson et al., 2017 <https://arxiv.org/abs/1704.03459>`_).
+For a discussion of ``dyPolyChord``'s performance, see the `performance section <http://dyPolyChord.readthedocs.io/en/latest/performance.html>`_ of the documentation.
 
 Documentation contents
 ----------------------
@@ -29,6 +32,7 @@ Documentation contents
    install
    demo
    api
+   performance
    likelihoods_and_priors
 
 Attribution
