### 2018-09-22 90f52ff
adding reference to performance tests
