### 2018-09-22 ejhigson/dyPolyChord#2 @ejhigson
Thanks for pointing out this requirement to me! There are empirical tests of dyPolyChord in the dynamic nested sampling paper which I think should be sufficient. If required, the code for reproducing these tests is publicly available at https://github.com/ejhigson/dns.

I have added both a reference to these results and a link to the code to reproduce them in the submitted paper. I have also added a performance section to the documentation which discusses this.

If you are happy with this I will close the issue.
