### 2018-09-22 ejhigson/dyPolyChord#1 @ejhigson
Again thanks for pointing this out! I have added this bit of the paper to the software documentation at https://dypolychord.readthedocs.io/en/latest/.

If you are happy with this I will close the issue.
