### 2018-09-24 ejhigson/dyPolyChord#1 @andrewfowlie
I think the docs are still missing

> who the target audience is?

and it could be much clearer about what it calculates, i.e. the Bayesian evidence. I think both should be in the first few (if not the first) paragraph of the docs.
