### 2018-09-24 ejhigson/dyPolyChord#3 @andrewfowlie
Functionality - Installation
Hi,

For my [review](https://github.com/openjournals/joss-reviews/issues/965), I am trying to check-off this one,

> ### Functionality
> - [ ] **Installation:** Does installation proceed as outlined in the documentation?

I have two problems.

1. It is very easy to install your package, run the tests, and still miss the fact that PolyChord and/or PyPolyChord are completely missing. I think this could be improved. E.g., since this is a Python package and PolyChord v1.14 is required, why not make PyPolyChord a requirement and indicate that it is missing?

2. Without PyPolyChord, I could run the tests quite quickly (and a few of them were skipped). Upon installing PyPolyChord, I ran the tests. I saw lots of `Polychord Warning: Non deterministic loglikelihood`, and it ran for a long time without finishing. A note about this warning and some indication about how long this step might take would be helpful.
