### 2018-09-24 ejhigson/dyPolyChord#2 @andrewfowlie
OK. @arfon Can you clarify whether the below requirement means  *the reviewer (me)* has confirmed the claims? or that *the authors* have confirmed the claims?

> - [ ] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

The authors have fairly thoroughly investigated the performance of their algorithm (and in fact the implementation submitted to JOSS) in a separate publication, but the submission to JOSS doesn't itself contain examples of the performance and I haven't verified their findings.

@ejhigson If I must verify the performance claims myself, if possible, could you prepare a minimal example that shows orders of magnitude gains in computational efficiency? Preferably, the example shouldn't require great deal of CPU time.
