### 2018-09-24 ejhigson/dyPolyChord#4 @andrewfowlie
Documentation - Functionality documentation
For my [review](https://github.com/openjournals/joss-reviews/issues/965), I am almost happy to check this requirement, as the API method documentation is generally really great.

> ### Documentation
> - [x] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I have two comments though:

1. The names/location of the output files should be made clearer. E.g., the [run_dynamic_ns docs](https://dypolychord.readthedocs.io/en/latest/api.html#module-dyPolyChord.run_dynamic_ns) don't give a clear indication of where the new results will be written and what they contain. A user unfamiliar with PolyChord etc might be confused.

2. The code in many places favors the syntax

```
def myfunction(x, y, *kwargs)
    z = kwargs.pop('z', 10)
   ...
```
to
```
def myfunction(x, y, z=10)
    ....
```
I don't understand the preference for that? and it makes reading the docs and code trickier for me as I can't tell much by just glancing at the signature. But I think there must be a reason? and it's not a big enough deal for me to insist that it's changed.
