### 2018-09-24 ejhigson/dyPolyChord#2 @ejhigson
> @ejhigson If I must verify the performance claims myself, if possible, could you prepare a minimal example that shows orders of magnitude gains in computational efficiency? Preferably, the example shouldn't require great deal of CPU time.

I am afraid running such a test (without a somewhat contrived problem or choice or priors) is computationally expensive as the big gains are in highish dimensions and computational cost with PolyChord scales with dimension D as ~D^3. Plus to do so rigorously you need a lot of runs to measure the speedup accurately and an analytic value for the result to check you have not reduced the variation of results by introducing biases.

For this reason in the JOSS paper I only claim dynamic nested sampling (in general) "allows order-of-magnitude increases in efficiency" and that dyPolyChord (this implementation) offers "state-of-the-art performance" - i.e. better than PolyChord. I think these claims are well verified in https://arxiv.org/abs/1704.03459, and I don't think you need to rerun the tests yourself (they are computationally expensive) although I of course defer to @arfon on this.
