### 2018-09-24 fd60a4f
added intro sentence to docs
