### 2018-09-24 ejhigson/dyPolyChord#1 @ejhigson
This is a good point - now looking at it again you are definitely right. I have added a sentence saying what nested sampling calculates as the first sentence of the homepage here https://dypolychord.readthedocs.io/en/latest/.

I am going to provisionally close this issue but let me know if you think it should be reopened!
