### 2018-09-24 ejhigson/dyPolyChord#1 @ejhigson
I should add: I think it is implicit that the target audience is "people who want to calculate Bayesian evidences and/or posterior samples from a likelihood and prior", but let me know if you think this should be further clarified.
