### 2018-09-24 bab2962
added mention of high-dimensional parameter estimation
