### 2018-09-24 ae3116c
docstring updates
