### 2018-09-24 ejhigson/dyPolyChord#4 @ejhigson
> 1. The names/location of the output files should be made clearer. E.g., the [run_dynamic_ns docs](https://dypolychord.readthedocs.io/en/latest/api.html#module-dyPolyChord.run_dynamic_ns) don't give a clear indication of where the new results will be written and what they contain. A user unfamiliar with PolyChord etc might be confused.

I have added this (in the docstring of run_dypolychord).
