### 2018-09-24 ejhigson/dyPolyChord#3 @ejhigson
> 1. It is very easy to install your package, run the tests, and still miss the fact that PolyChord and/or PyPolyChord are completely missing. I think this could be improved. E.g., since this is a Python package and PolyChord v1.14 is required, why not make PyPolyChord a requirement and indicate that it is missing?

The tests are deliberately designed to run without PyPolyChord so people who only use C++ and Fortran likelihoods can perform tests without installing PyPolyChord. Similarly the tests check the code required for C++ and Fortran likelihoods without needing a compiled executable (which would be a pain for the user). Another reason for this approach is that it allows the use of continuous integration with travis CI, which would otherwise be very difficult or impossible due to the need to download PolyChord from sourceforge with a login.

You are correct that this is a bit confusing - I have added a warning to the tests which triggers if PyPolyChord is not installed, reminding the user that it isnt installed but is only needed if they want to use Python likelihoods.
