### 2018-09-24 ejhigson/dyPolyChord#3 @ejhigson
> 2\. Without PyPolyChord, I could run the tests quite quickly (and a few of them were skipped). Upon installing PyPolyChord, I ran the tests. I saw lots of `Polychord Warning: Non deterministic loglikelihood`, and it ran for a long time without finishing. A note about this warning and some indication about how long this step might take would be helpful.

This is not intended to happen! What version of PolyChord and Python are you using? And do you know which test this occurs in?

Do you get any warnings saying something like "This indicates that your PolyChord install\'s random number  generator is different to the one used to calculate the expected results, so I will skip tests of exact numerical values. If all the other tests pass then everything should work ok.'?

The tests use very low nlive and num_repeat settings to make them fast, which can cause this problem. However they also use random seeding which should prevent it from occurring. I suspect the seeding dosn't work on your machine and hence the tests need to be changed.
