### 2018-09-24 6ba7ec0
adding warning for tests using PyPolyChord
