### 2018-09-24 ejhigson/dyPolyChord#3 @andrewfowlie
Thanks for your responses. I’m happy that the first one is now addressed. About the second one, I’ll have to get back to you tomorrow after running the tests again. I’ll try to isolate which test caused that behaviour, whether other messages were printed, and tell you my version numbers etc.
