### 2018-09-24 ejhigson/dyPolyChord#3 @andrewfowlie
Sorry, I spoke too soon. I just checked the change https://github.com/ejhigson/dyPolyChord/commit/6ba7ec0a02e645ff7b3e17890e8785f2b24de417. It looks good, except that the warning could be printed if PyPolyChord is present but mpi4py isn’t.
