### 2018-09-24 ejhigson/dyPolyChord#2 @andrewfowlie
I think I misunderstood the software paper as claiming orders of magnitude improvement by your code. Upon re-reading it, I now see it doesn’t say that. I’m happy to accept the arXiv paper as a demonstration of the weaker claim of state-of-the-art performance.
