### 2018-09-24 fc4a90c
test update to seperate import errors from PyPolyChord and mpi4py
