### 2018-09-24 ejhigson/dyPolyChord#3 @ejhigson
Well spotted! Fix in fc4a90ce51c4fd1affab284671b9393f97401081
