### 2018-09-25 ejhigson/dyPolyChord#3 @andrewfowlie
# Problem

After git pulling today (latest commit fc4a90ce51c4fd1affab284671b9393f97401081), I found the same as yesterday,
```
andrew@andrew-laptop:~/repos/dyPolyChord$ nosetests &> nosetests.log
andrew@andrew-laptop:~/repos/dyPolyChord$ more nosetests.log
............................F/home/andrew/repos/dyPolyChord/dyPolyChord/nlive_allocation.py:49: UserWarning: Smoothing has added turning points to nlive allocation when dynamic_goal=0. I am using the unsmoothed nlives instead.
  'instead.'), UserWarning)
Polychord Warning: Non deterministic loglikelihood
Polychord Warning: Non deterministic loglikelihood
Polychord Warning: Non deterministic loglikelihood
```
The warning is repeated many times and it doesn't finish in a reasonable time.

# Version numbers etc

My versions are
```
python -V
Python 2.7.15rc1

>>> numpy.__version__
'1.15.1'
>>> nestcheck.__version__
'0.1.8'
>>> scipy.__version__
'1.1.0'

PolyChord_v1.15.tar.gz
```
The installations all went OK, with the exception of PyPolyChord, where I had to `export LD_PRELOAD=/usr/lib/x86_64-linux-gnu/libmpi_cxx.so` to avoid a missing symbol problem with MPI. I believe though that my installation is fine, though unfortunately there aren't any tests, e.g.,
```
andrew@andrew-laptop:~/codes/PolyChord$ python run_PyPolyChord.py
...
 ____________________________________________________
|                                                    |
| ndead  =         1039                              |
| log(Z) =           -1.73781 +/-            0.24593 |
|____________________________________________________|
Install matplotlib and getdist for plotting examples

```

# Isolating tests that fail

By investigating further I found that the test that doesn't finish in a reasonable time is `test_dynamic_evidence`. When I remove that one by adding `return True`
```
andrew@andrew-laptop:~/repos/dyPolyChord$ git diff
diff --git a/tests/tests.py b/tests/tests.py
index 8067f7f..4194899 100644
--- a/tests/tests.py
+++ b/tests/tests.py
@@ -128,6 +128,7 @@ class TestRunDyPolyChordNumers(unittest.TestCase):

     def test_dynamic_evidence(self):
         """Test numerical results for nested sampling with dynamic_goal=0."""
+        return True
         dynamic_goal = 0
         dyPolyChord.run_dypolychord(
             self.run_func, dynamic_goal, self.settings,
```
I find though that two other tests fail
```
andrew@andrew-laptop:~/repos/dyPolyChord$ nosetests
............................F.F.....
FAIL: Test numerical results for nested sampling with
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 163, in test_dynamic_both_evidence_and_param
    self.assertEqual(e.count_samples(run), 556)
AssertionError: 426 != 556

FAIL: Test numerical results for nested sampling with dynamic_goal=1.
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 182, in test_dynamic_param
    places=12)
AssertionError: 0.06839158405509836 != 0.029776073688009493 within 12 places

Ran 36 tests in 4.719s

FAILED (failures=2)
```

# Tweaking `precision_criterion`

i reduced the `precision_criterion` as your code indicates that this could be related to the non-deterministic errors.

```
andrew@andrew-laptop:~/repos/dyPolyChord$ git diff
diff --git a/tests/tests.py b/tests/tests.py
index 8067f7f..c7d3513 100644
--- a/tests/tests.py
+++ b/tests/tests.py
@@ -111,7 +111,7 @@ class TestRunDyPolyChordNumers(unittest.TestCase):
             # errors. These occur due in the low dimension and low and nlive
             # cases we use for fast testing as runs sometimes get very close
             # to the peak where the likelihood becomes approximately constant.
-            'precision_criterion': 0.01,
+            'precision_criterion': 0.001,
             'seed': 1,
             'max_ndead': -1,
             'base_dir': TEST_CACHE_DIR,
```
I find that it then runs alright, but still fails
```
andrew@andrew-laptop:~/repos/dyPolyChord$ nosetests
............................F/home/andrew/repos/dyPolyChord/dyPolyChord/nlive_allocation.py:49: UserWarning: Smoothing has added turning points to nlive allocation when dynamic_goal=0. I am using the unsmoothed nlives instead.
  'instead.'), UserWarning)
FF.....
FAIL: Test numerical results for nested sampling with
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 162, in test_dynamic_both_evidence_and_param
    self.assertEqual(e.count_samples(run), 556)
AssertionError: 525 != 556

FAIL: Test numerical results for nested sampling with dynamic_goal=0.
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 143, in test_dynamic_evidence
    self.assertEqual(e.count_samples(run), 548)
AssertionError: 554 != 548

FAIL: Test numerical results for nested sampling with dynamic_goal=1.
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 181, in test_dynamic_param
    places=12)
AssertionError: 0.09796131954595269 != 0.029776073688009493 within 12 places

Ran 36 tests in 5.565s

FAILED (failures=3)
```
