### 2018-09-25 70a9e62
updated test settings to imporve robustness to non-deterministic likelihood problems which random seeding does not work as expected
