### 2018-09-25 ejhigson/dyPolyChord#3 @ejhigson
Thank you very much for this!!

I think the problem is due to the previous settings not being very robust when the random seeing does not give them exactly the result expected. I have changed the settings in a manner which should solve this problem in 70a9e62f5033665b4c2ef936c7643fd7c8fad517, and updated the values the tests expect to the ones I get on my machine (the tests checking the result values caused the failures you show above). The tests are a little slower but they still don't take long so I think this is fine.

Let me know if this works for you!
