### 2018-09-26 ejhigson/dyPolyChord#3 @andrewfowlie
```
andrew@andrew-laptop:~/repos/dyPolyChord$ git pull
Updating fc4a90c..70a9e62
Fast-forward
 tests/tests.py | 34 ++++++++++++++++++----------------
 1 file changed, 18 insertions(+), 16 deletions(-)
andrew@andrew-laptop:~/repos/dyPolyChord$ export LD_PRELOAD=/usr/lib/x86_64-linux-gnu/libmpi_cxx.so
andrew@andrew-laptop:~/repos/dyPolyChord$ nosetests
............................FFF.....
FAIL: Test numerical results for nested sampling with
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 164, in test_dynamic_both_evidence_and_param
    self.assertEqual(e.count_samples(run), 1093)
AssertionError: 920 != 1093

FAIL: Test numerical results for nested sampling with dynamic_goal=0.
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 145, in test_dynamic_evidence
    self.assertEqual(e.count_samples(run), 1169)
AssertionError: 924 != 1169

FAIL: Test numerical results for nested sampling with dynamic_goal=1.
Traceback (most recent call last):
  File "/home/andrew/repos/dyPolyChord/tests/tests.py", line 183, in test_dynamic_param
    places=12)
AssertionError: -0.029540931478677097 != -0.05323120028149568 within 12 places

Ran 36 tests in 33.492s

FAILED (failures=3)

```
