### 2018-09-27 84ee43a
fix for python2 float division error
