### 2018-09-27 ejhigson/dyPolyChord#3 @ejhigson
Gotcha! The random seeding is working ok - the problem was a rogue division of integers when computing how many samples to use

`samp_tot = init_run['logl'].shape[0] * (nlive_const / ninit)`

With a fresh download of the latest PolyChord, running in python 2, the fix in 84ee43ac91e2ddd3f88bd92847de87d202d8398d is now working for me whereas without the fix I got the exact same errors as you. I work almost exclusively in python 3 and this must have somehow slipped past my previous python2 testing.

Thank you very much for your help finding this @andrewfowlie! If you confirm this is working for you I will close the issue
