### 2018-09-27 ejhigson/dyPolyChord#3 @andrewfowlie
```
andrew@andrew-laptop:~/repos/dyPolyChord$ nosetests
....................................
Ran 36 tests in 28.845s

OK
```
Great! Thanks for investigating this.
