### 2018-09-27 #965 @andrewfowlie
I have completed the checklist of tasks. The program, documentation and software paper were well-written and I think the program could be very useful for the scientific community.
