### 2018-09-27 #965 @ejhigson
@zhampel thank you very much for taking the time to review this software - I really appreciate it! Do you know how long it will take you to finish your review?

I am actually going to submit my PhD thesis on Monday and it would be ideal if this is all done by then and I can cite the paper in it, but this is no big deal so no worries at all if that is not convenient!
