### 2018-09-27 657beb8
very minor edit to intro sentence
