### 2018-09-27 a079c84
Minor reference typo fixes
