+++ docs/paper/paper.bib
@@ -107,7 +107,7 @@ author = {Higson, Edward and Handley, Will and Hobson, Mike and Lasenby, Anthony
 eprint = {1704.03459},
 journal = {arXiv preprint arXiv:1704.03459},
 keywords = {bayesian evidence,nested sampling,parameter estimation},
-title = {{Dynamic nested sampling : an improved algorithm for parameter estimation and evidence calculation}},
+title = {{Dynamic nested sampling: an improved algorithm for parameter estimation and evidence calculation}},
 url = {https://arxiv.org/abs/1704.03459},
 year = {2017}
 }
@@ -134,7 +134,7 @@ doi = {10.21105/joss.00916},
 journal = {Journal of Open Source Software},
 number = {29},
 pages = {916},
-title = {nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations},
+title = {{nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations}},
 volume = {3},
 year = {2018}
 }
