### 2018-09-29 #965 @zhampel
@ejhigson Yes, I have been checking things for the last few days. I wanted to make a few notes about Installation, Performance and Tests. I will not require these for publication, but I will require you make a few notes in the documentation regarding a few comments I have, and I **strongly** suggest you address after your thesis ;)
