### 2018-09-29 #965 @zhampel
Installation
========

Python
-------
I had no troubles installing and subsequently running with Python 3. Your setup.py script works seamlessly there.

However, I did have issues with Python 2.7. The steps required to get running in Python 2.7 were the following:
1. `virtualenv -p /usr/bin/python2 dyPolyChord_review`
2. `source dyPolyChord_review/bin/activate`
3. `python setup.py install`
4. `pip install futures pytz python-dateutil`
If I didn't perform step 4, then I would get the following main error messages (including a bunch of other import garbage) when running `nosetests`:
- nose.plugins.cover: ERROR: Coverage not available: unable to import coverage module
EE
ERROR: Failure: ImportError (No module named concurrent.futures)
- ImportError: No module named concurrent.futures
- ImportError: Missing required dependencies ['pytz', 'dateutil']

After doing so, then the `nosetests` worked fine.

I suggest you simply make a note in the docs that in Python 2.7 you need to run step 4 OR you take time to address this in your setup script reqs. But again, given your thesis, I think the docs method is  sufficient. Of course, if you haven't had these issues before, then please explicitly provide some cmd line steps that work for you.

Finally, with Python 2.7 on my Mac, I had to run additionally run the following to get the `python setup.py install` step to work:
- `pip install pytest-runner nestcheck`

Installation & Performance
===================
PyPolyChord
-------------
In order to get the PyPolyChord functionality to work, I had to run `sudo apt-get install libblacs-mpi-dev` to get mpif90 installed. There is not streamlined PolyChord installation documentation that I could see in the package, unless I completely missed something. In addition, I see that *CCPForge is closing*, so there another repo must be made available for its use in the near future. This is of course beyond the scope of this review, but I wanted to make note of it.

I know that you've discussed and addressed this with @andrewfowlie, so I understand that the package is designed to work without PyPolyChord, but I think that could be clearer in the docs in one of two ways:
- Make the fact that the tests can run without PyPolyChord uber clear, and perhaps provide some examples in the Performance section instead of just pointing the user to your dns package (perhaps you have some figures from writing your thesis... ;) )
- Provide a streamlined installation procedure for all parts, so that the Performance can be verified in a straightforward manner.

If you can show a few examples that would be best, but I leave that up to you.
