### 2018-09-29 #965 @zhampel
@ejhigson I think you should make a few notes in your documentation addressing my comments. I have however, checked off all the boxes because the dyPolyChord project is well written, does work as advertised, and is well documented. Good luck with your thesis!

For the purposes of this submission, @arfon I have completed my review.
