### 2018-09-30 0564282
doc updates on tests not needing dyPolyChord
