+++ docs/install.rst
@@ -33,6 +33,7 @@ Dependencies
 
 
 ``PolyChord`` is available at https://ccpforge.cse.rl.ac.uk/gf/project/polychord/ and has its own installation and licence instructions; see the link for more information.
+Note running ``dyPolyChord`` with compiled C++ or Fortran likelihoods does not require the installation of ``PyPolyChord`` (``PolyChord``'s Python interface).
 
 
 Tests
@@ -50,4 +51,5 @@ To also get code coverage information (this requires the ``coverage`` package), 
 
     nosetests --with-coverage --cover-erase --cover-package=dyPolyChord
 
-If all the tests pass, the install should be working.
+Note that these tests will run without ``PolyChord``. This is to allow all the ``dyPolyChord`` code (including code specifically for Python or compiled likelihoods) to be tested without the need for the user to compile any executables or install ``PyPolyChord``.
+If ``PyPolyChord`` is installed, the tests will also run calculations using Python likelihoods and check their results (otherwise these tests are skipped).
