### 2018-09-30 #965 @ejhigson
> @ejhigson Yes, I have been checking things for the last few days. I wanted to make a few notes about Installation and Performance. I will not require these for publication, but I will require you make a few notes in the documentation regarding a few comments I have, and I **strongly** suggest you address after your thesis ;)

Thank you so much for being so accommodating!!

> However, I did have issues with Python 2.7. The steps required to get running in Python 2.7 were the following:
>
> 1. `virtualenv -p /usr/bin/python2 dyPolyChord_review`
> 2. `source dyPolyChord_review/bin/activate`
> 3. `python setup.py install`
> 4. `pip install futures pytz python-dateutil`
>    If I didn't perform step 4, then I would get the following main error messages (including a bunch of other import garbage) when running `nosetests`:
>
> * nose.plugins.cover: ERROR: Coverage not available: unable to import coverage module
>   EE
>   ======================================================================
>   ERROR: Failure: ImportError (No module named concurrent.futures)
> * ImportError: No module named concurrent.futures
> * ImportError: Missing required dependencies ['pytz', 'dateutil']
>
> After doing so, then the `nosetests` worked fine.
>
> I suggest you simply make a note in the docs that in Python 2.7 you need to run step 4 OR you take time to address this in your setup script reqs. But again, given your thesis, I think the docs method is sufficient. Of course, if you haven't had these issues before, then please explicitly provide some cmd line steps that work for you.

Thank you for pointing this out! I like your testing with a fresh virtualenv - I will use this in future. I managed to replicate this issue and fix it in 05642826781b73868a04f6ba629ce5d7fb271188 by adding
`
':python_version == "2.7"': ['futures', 'pytz',  'python-dateutil']`

to setup.py.
