### 2018-09-30 #965 @ejhigson
> ## PyPolyChord
> In order to get the PyPolyChord functionality to work, I had to run `sudo apt-get install libblacs-mpi-dev` to get mpif90 installed. There is not streamlined PolyChord installation documentation that I could see in the package, unless I completely missed something. In addition, I see that _CCPForge is closing_, so there another repo must be made available for its use in the near future. This is of course beyond the scope of this review, but I wanted to make note of it.
>
> I know that you've discussed and addressed this with @andrewfowlie, so I understand that the package is designed to work without PyPolyChord, but I think that could be clearer in the docs in one of two ways:
>
> * Make the fact that the tests can run without PyPolyChord uber clear, and perhaps provide some examples in the Performance section instead of just pointing the user to your dns package (perhaps you have some figures from writing your thesis... ;) )
> * Provide a streamlined installation procedure for all parts, so that the Performance can be verified in a straightforward manner.
>
> If you can show a few examples that would be best, but I leave that up to you.

Thank you for pointing this out! Unfortunately I don't have any control over PolyChord and its docs, install method etc. but I know the PolyChord authors are working on improved versions and are aware of the CCPForge closure.

I have updated the dyPolyChord install docs to make it (hopefully) much clearer that the tests run without PolyChord and why they do this. I also do indeed have more tests/plots of dyPolyChord performance which will be released after my PhD :) Unfortunately rigorous performance verification is rather computationally expensive since, the gains are biggest in high dimensional problems and the calculations need to be repeated many times. Nevertheless I intend to make my tests reproducible as much as possible given practical limitations.
