### 2018-09-30 #965 @ejhigson
@andrewfowlie and @zhampel thank you very much for taking the time to review this submission! Your feedback has been extremely helpful!

@arfon I think the reviews are now complete. I have archived an up to date version on zenodo at 10.5281/zenodo.1439193
